+++ core/src/tests/test_config_parser_sd.cc
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2019-2023 Bareos GmbH & Co. KG
+   Copyright (C) 2019-2024 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -36,8 +36,8 @@ TEST(ConfigParser_SD, test_stored_config)
 {
   OSDependentInit();
 
-  std::string path_to_config_file = std::string(
-      RELATIVE_PROJECT_SOURCE_DIR "/configs/bareos-configparser-tests");
+  std::string path_to_config_file
+      = std::string("configs/bareos-configparser-tests");
   my_config = InitSdConfig(path_to_config_file.c_str(), M_CONFIG_ERROR);
   ParseSdConfig(configfile, M_CONFIG_ERROR);
 
@@ -70,9 +70,8 @@ TEST(ConfigParser_SD, CFG_TYPE_STR_VECTOR_OF_DIRS)
   InitMsg(nullptr, nullptr);
 
   std::string path_to_config_file
-      = std::string(RELATIVE_PROJECT_SOURCE_DIR
-                    "/configs/bareos-configparser-tests/bareos-sd-")
-        + test_name + std::string(".conf");
+      = std::string("configs/bareos-configparser-tests/bareos-sd-") + test_name
+        + std::string(".conf");
   my_config = InitSdConfig(path_to_config_file.c_str(), M_ERROR_TERM);
   ASSERT_TRUE(my_config->ParseConfig());